requirements.txt
accelerate==1.10.1
altair==5.5.0
altgraph @ file:///AppleInternal/Library/BuildRoots/39d9dc1a-2111-11f0-be06-226177e5bb69/Library/Caches/com.apple.xbs/Sources/python3/altgraph-0.17.2-py2.py3-none-any.whl
annotated-doc==0.0.3
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.11.0
appnope==0.1.4
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.4.0
asttokens==3.0.0
async-lru==2.0.5
attrs==25.4.0
babel==2.17.0
beautifulsoup4==4.14.2
bleach==6.2.0
blinker==1.9.0
bottle==0.13.4
cachetools==6.2.2
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.1.8
colorlog==6.10.1
comm==0.2.3
debugpy==1.8.17
decorator==5.2.1
defusedxml==0.7.1
dill==0.4.0
docling==2.61.2
docling-core==2.51.0
docling-ibm-models==3.10.2
docling-parse==4.7.1
et_xmlfile==2.0.0
exceptiongroup==1.3.0
executing==2.2.1
Faker==37.12.0
fastapi==0.121.0
fastjsonschema==2.21.2
filelock==3.19.1
filetype==1.2.0
fqdn==1.5.1
fsspec==2025.10.0
future @ file:///AppleInternal/Library/BuildRoots/39d9dc1a-2111-11f0-be06-226177e5bb69/Library/Caches/com.apple.xbs/Sources/python3/future-0.18.2-py3-none-any.whl
gitdb==4.0.12
GitPython==3.1.45
grpcio==1.76.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
idna==3.11
importlib_metadata==8.7.0
ipykernel==6.31.0
ipython==8.18.1
ipywidgets==8.1.8
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.2
json5==0.12.1
jsonlines==4.0.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonref==1.1.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyter_server==2.17.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.10
jupyterlab_pygments==0.3.0
jupyterlab_server==2.28.0
jupyterlab_widgets==3.0.16
langchain-core==0.3.79
langchain-docling==1.1.0
langchain-huggingface==0.3.1
langchain-milvus==0.2.1
langchain-ollama==0.3.10
langchain-text-splitters==0.3.11
langsmith==0.4.37
lark==1.3.1
latex2mathml==3.78.1
lxml==6.0.2
macholib @ file:///AppleInternal/Library/BuildRoots/39d9dc1a-2111-11f0-be06-226177e5bb69/Library/Caches/com.apple.xbs/Sources/python3/macholib-1.15.2-py2.py3-none-any.whl
markdown-it-py==3.0.0
marko==2.2.1
MarkupSafe==3.0.3
matplotlib-inline==0.2.1
mdurl==0.1.2
milvus==2.3.9
milvus-lite==2.5.1
mistune==3.1.4
mpire==2.10.2
mpmath==1.3.0
multiprocess==0.70.18
narwhals==2.11.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.2.1
notebook==7.4.7
notebook_shim==0.2.4
numpy==2.0.2
ocrmac==1.0.0
ollama==0.6.1
omegaconf==2.3.0
opencv-python==4.12.0.88
openpyxl==3.1.5
orjson==3.11.4
overrides==7.7.0
packaging==25.0
pandas==2.3.3
pandocfilters==1.5.1
parso==0.8.5
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.4.0
pluggy==1.6.0
polyfactory==2.22.4
prometheus_client==0.23.1
prompt_toolkit==3.0.52
protobuf==6.33.1
proxy_tools==0.1.0
psutil==7.1.3
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==21.0.0
pyclipper==1.3.0.post6
pycparser==2.23
pydantic==2.12.4
pydantic-settings==2.11.0
pydantic_core==2.41.5
pydeck==0.9.1
Pygments==2.19.2
pyinstaller==6.16.0
pyinstaller-hooks-contrib==2025.9
pylatexenc==2.10
pymilvus==2.6.3
pyobjc-core==12.0
pyobjc-framework-Cocoa==12.0
pyobjc-framework-CoreML==12.0
pyobjc-framework-Quartz==12.0
pyobjc-framework-Security==12.0
pyobjc-framework-UniformTypeIdentifiers==12.0
pyobjc-framework-Vision==12.0
pyobjc-framework-WebKit==12.0
pypdfium2==4.30.0
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.2.1
python-json-logger==4.0.0
python-pptx==1.0.2
pytz==2025.2
pywebview==6.1
PyYAML==6.0.3
pyzmq==27.1.0
rapidocr==3.4.2
referencing==0.36.2
regex==2025.11.3
requests==2.32.5
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rich==14.2.0
rpds-py==0.27.1
rtree==1.4.1
safetensors==0.6.2
scikit-learn==1.6.1
scipy==1.13.1
semchunk==2.2.2
Send2Trash==1.8.3
sentence-transformers==5.1.2
shapely==2.0.7
shellingham==1.5.4
six @ file:///AppleInternal/Library/BuildRoots/39d9dc1a-2111-11f0-be06-226177e5bb69/Library/Caches/com.apple.xbs/Sources/python3/six-1.15.0-py2.py3-none-any.whl
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.8
stack-data==0.6.3
starlette==0.49.3
streamlit==1.50.0
sympy==1.14.0
tabulate==0.9.0
tenacity==9.1.2
terminado==0.18.1
threadpoolctl==3.6.0
tinycss2==1.4.0
tokenizers==0.22.1
toml==0.10.2
tomli==2.3.0
torch==2.8.0
torchvision==0.23.0
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.57.1
tree-sitter==0.23.2
tree-sitter-c==0.23.6
tree-sitter-java==0.23.5
tree-sitter-javascript==0.23.1
tree-sitter-python==0.23.6
tree-sitter-typescript==0.23.2
typer==0.19.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uri-template==1.3.0
urllib3==2.5.0
uvicorn==0.38.0
wcwidth==0.2.14
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.9.0
widgetsnbextension==4.0.15
xlsxwriter==3.2.9
zipp==3.23.0
zstandard==0.25.0
